rename: Ns3Log.* -> Log.*
